Changes in the RtVMap scaled data retrieving capability and added ScaleSignal method for devices

+++ Source/Components/DataSources/UEIAI217DataSource/UEIMapContainer.cpp
@@ -331,7 +331,7 @@ bool UEIMapContainer::StartMap(){
     return false;
 }
 
-bool UEIMapContainer::PollForNewPacket(float64* destinationAddr){
+bool UEIMapContainer::PollForNewPacket(float32* destinationAddr){
     return false;
 }
